lambda-functions-gcp-find-text/requirements.txt
google-cloud-vision==1.0.0
google-cloud-storage==1.28.1
google-cloud-datastore==1.12.0
google-cloud==0.34.0
google-cloud-pubsub==1.5.0
